Final? Version of Analytics Logging
AnaSetup was updated to simplify logging process. Logging is consolidated into one function - logEvent(String Action).
Simply provide a string as a tag for whatever you wish to record in the log.
AnaViewer received a complete update - it now reads through the log file, counting up the number of times each event happens. It will record to and update to a report<date>.txt file for each day of the year. To start this process, call processLogs(). This should be done just before generating a graph report and just before closing the program.
Both of these types of files are located at C:\\AnaManager